add optimizations

+++ src/forest_property/tree.rs
@@ -1,12 +1,12 @@
-#[derive(Default, Debug, Clone)]
+#[derive(Default, Debug, Clone, Copy)]
 pub struct Tree {
-    species: i64,
-    mean_height: f64,
+    species: u8,
+    mean_height: f32,
     position: (f64, f64, f64),
 }
 
 impl Tree {
-    pub fn new(species: i64, mean_height: f64, position: (f64, f64, f64)) -> Self {
+    pub fn new(species: u8, mean_height: f32, position: (f64, f64, f64)) -> Self {
         Tree {
             species,
             mean_height,
@@ -14,7 +14,7 @@ impl Tree {
         }
     }
 
-    pub fn species(&self) -> i64 {
+    pub fn species(&self) -> u8 {
         self.species
     }
